requirements.txt
Django>=5.1.3
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
drf-spectacular>=0.27.0
djangorestframework-simplejwt>=5.3.0
aiofiles
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.4.0
django-cors-headers
celery
redis
django-celery-beat
django-prometheus
boto3
django-storages

pytz

paramiko
scp
loguru

firebase-admin